Re-add point.score in the search results
Going through the Getting Started guide, I noticed calling the /search
endpoint returned results with `0.0` for the `confidence_score`, which
didn't match the guide contents. It looks like a static value was
introduced in [here in

I'm not sure why it would be intentional, but in case it's not, this PR
re-introduces `point.score` to the output.