implement storing metadata alongside vector
Add metadata to add_embedding() arguments, store it together with the
vector.

Add get_points() and update_metadata() functions to VectorDb to retrieve
metadata and update it.

When creating content for the first time, send metadata together with
the embedding.

If updating existing content, retrieve existing metadata and combine it
with the new one. If embedding is also updated, send updated metadata
with add_embedding(). If only metadata is updated, use update_metadata()
to change existing vector.