implement storing metadata alongside vector (#477)
* implement storing metadata alongside vector

Add metadata to add_embedding() arguments, store it together with the
vector.

Add get_points() and update_metadata() functions to VectorDb to retrieve
metadata and update it.

When creating content for the first time, send metadata together with
the embedding.

If updating existing content, retrieve existing metadata and combine it
with the new one. If embedding is also updated, send updated metadata
with add_embedding(). If only metadata is updated, use update_metadata()
to change existing vector.

* remove storing content id in qdrant metadata

Do not need to store content id in metadata, it can be retrieved from
point id.

* add tests

handle the case when there is existing metadata and an embedding is
added.

* add implementation of get_points and update_metadata for pg_vector

* implement storing metadata in lancedb

Switch to crate lancedb from vectordb

Store metadata json keys as columns in lancedb.

Update schema when new keys are used in metadata.

* pass list of tables content may be present in to ingest tasks

When creating an extraction task determine the list of tables content
may be present and pass it to extractor. When updating metadata for
existing content try to update all potentially affected tables with new
metadata.

* comment Task fields

Reduce number of calls to combine_metadata() and number of combine functions.