Fix broken tests and delete function image tests
We didn't run these tests in CD for a while so some where broken.
Delete function image tests because this functionality is going
to get reworked. All of them are hard failing now and it'll take
a lot of time to fix them.

Now `make test` is passing.

I also made run_tests.sh command more conveniet. It fails if INDEXIFY_URL
is not defined and adds a line stating status of the test run. More
convenient to understand status of tests for people because the tests
output a lot of text.

I'll add these tests to our GitHub actions in an upcoming PR.