Use python-sdk `make test` target to run all python-sdk tests
`make test` command now provides full official set of unit and
integration tests for python-sdk. They can be started using the
command during development and in CI. This is why PR template is
updated with the command and GitHub CI actions run it now.

Also use `make build` command instead of calling `poetry install`
exlicitly in CI because `make build` is a source of truth command
for building of python-sdk (it runs `poetry install` too).