Host Executor runs Functions Executor processes to run customer functions
Each task gets its own Function Executor process.

All file caching is removed from Executor. Every task gets its
own directory with its own copy of serialized Graph and function
inputs and outputs. This is done to simplify the implementation
for now and also to fix the bug when we delete a graph and then
create a new one with the same name. In this case Executor still
uses the deleted graphs' data because the caching key uses graph
name and we don't have graph IDs yet. The removal of caching fixes
this issue.

Also removed redundant num_workers configuration and some other
dead code. This configuration is not used and is now completely
obsolete with the Function Executor in separate processes.

Testing:

make test
make fmt
make check