Terminate Indexify processes in the end of GH Action test runs
Currently we don't cleanely terminate Indexify Server and Executor
in the end of test runs. This results in a misleading collection
of many orphaned processes in the end of test runs. Do the clean
termination to remove the misleading orphaned processes collection
and to run the code in Executor that terminates Function Executors
cleanely. This ensures better line coverage in our CI GH Actions
testing.