Function allowlist for Executor and manual Graph management (#1146)
* Function allowlist for Executor

Making the executors only run functions of specified compute graphs.
The functions are specified as

`indexify-cli executor --function <namespace>:<workflow>:<compute_func>:<version> --function ...`.

The allowlist is used with Executors in production setups. E.g. only run
a certain function in a Kubernetes POD with an Executor.
This is a default production setup because each function has different
resource usage and it’s easier to size containers per function.
Same with other container configs like secrets, roles, volumes.

`indexify-cli executor --dev` mode is still present. It makes server try
to run any function on the executor. This is convenient for development
of Indexify and for integration testing.

Co-authored-by: Diptanu Gon Choudhury <diptanuc@gmail.com>
Co-authored-by: Eugene Batalov <eugene@tensorlake.ai>

* Change Graph version type from u32 to string and set it in SDK

This allows users to implement versioning of their graphs with
any semantic they want. Use a random uuid by default for the
graph versions so if a user doesn't want to manage versions
manually we just always update the graph on each RemoteGraph.deploy().

---------

Co-authored-by: Eugene Batalov <eugene@tensorlake.ai>