Extend FunctionExecutorServerConfiguration with graph metadata (#1361)
The new image builder implementation centralizes the knowledge of image URIs into the image builder V2 service. This is due to security constraints implemented for pulling images from our container registry.

The signature to pull images in this new service requires the usage of the namespace, graph name, graph version, and function name.

At the moment, these are not accessible to services looking to implement a FunctionExecutorServerFactory that relies on those values.

This PR extends FunctionExecutorServerConfiguration to hold those values.