Groundwork for Executor resource reporting and GPU allocation
Change GPUModel gRPC enum into SDK GPU Model strings.
This way we won't need to convert values back and forth. We'll use
the same GPU model strings across SDK, Server and Executor.

Fetch NVIDIA GPU info on Executor startup and report it to server.
Simple GPU allocator for functions. It won't be used in OSS but it's
needed for Executor free resources reporting which is in OSS too.