Use gRPC proto enums for GPU models
We can't hardcode GPU models in Tensorlake SDK because they depend
on Indexify cluster. Because of this we don't have a single source
of truth for all supported GPU model values. Because of this we're
reverting back to a clear proto enum interface for GPU model values.
Server the will need to match its gpu model string values it gets
from SDK clients with the proto enum values we set on executor side.