Remove tests that make assertions on complex timings inside Indexify
As Executor is now pushing its state periodically to Server we can't rely
on sequences of events happening within a certain strict-ish duration
so we can't make assertions on these durations anymore.
Removing the tests that do such assertions for now until we have a better
approach to implement such tests.