Fix tests/cli/test_function_allowlist.py and refactor it
I changed the test to not run invocations sequentially but run them in parallel.
This is because when running sequentially there's effectively 1 task at a time
available to get distributed among 2 idle executors and it's not important to
route this task randomly to the executors. After making it parallel we have
1200 enqued tasks and we get really good uniform distribution in the test
and it's passing now.