feat: get desired state implementation (#1362)
* feat: get desired state implementation

* support all fields and change how in-mem is updated

* test: add more reconciliation tests

* fix: executor stop using clock as idempotency token

* fix: fix tests

* fix: fix rebase and make it possible to run server against minio/localstack

* feat: default to always do grpc reconciliation on executor

* fix behavior test

* lint

* fix tests params

* update test runner

* Update tensorlake

* Reduce periodic FE health checks period 10 sec -> 5 sec

This improves time to detect FE failures.

* Remove tests that make assertions on complex timings inside Indexify

As Executor is now pushing its state periodically to Server we can't rely
on sequences of events happening within a certain strict-ish duration
so we can't make assertions on these durations anymore.
Removing the tests that do such assertions for now until we have a better
approach to implement such tests.

* Fix tests/cli/test_function_allowlist.py and refactor it

I changed the test to not run invocations sequentially but run them in parallel.
This is because when running sequentially there's effectively 1 task at a time
available to get distributed among 2 idle executors and it's not important to
route this task randomly to the executors. After making it parallel we have
1200 enqued tasks and we get really good uniform distribution in the test
and it's passing now.

* Fix test_invoke_duration tests

The main idea is to not measure the extra 5 sec latency of Server learning
from Executor that Function Executor was created successfully.

* feat: support FE statup failures

* updated tensorlake commit

---------

Co-authored-by: Diptanu Gon Choudhury <diptanuc@gmail.com>
Co-authored-by: Eugene Batalov <eugene@tensorlake.ai>