Log Function Executor creation customer error message on Executor side
With direct BLOB store writes available for Executor it can now write
stdout, stderr files for each task that didn't run due to FE creation
failure. Executor should be doing all task logs writes anyway so we
shouldn't make exceptions for FE creation logs. This also allows to
remove this unnecessary responsibility from Server.

We'll FE creation logs BLOB store writes on Executor side later when
we remove a lot of legacy SSE stream code and rework state reconciler
design (possbily when working on batching of tasks). Write now it doesn't
make sense to implement it because the current architecture make it very
complex and because we have a lot of legacy code for SSE stream that we'd
delete soon.