Bring back --enable-grpc-state-reconciler Executor CLI flag (False by default)
By default this flag is False which means that we'll use SSE stream
by default. We need to be able to use the legacy SSE stream mode to
run with the stable published Server that doesn't use gRPC yet and
still uses SSE stream mode.

--enable-grpc-state-reconciler was originally removed in
https://github.com/tensorlakeai/indexify/pull/1362/files#diff-e207854b655720e9130572ef90eb931ffe8460c318ddd8f3c916a8756a29c066
we're bringing it back here.