feat: CLI target for Cloud OpenAPI spec generation (#1413)
## Context

For Tensorlake Cloud development, we require a different set of RESTful paths that comply with how the Tensorlake Ingress enriches requests.

This PR does not affect production, OSS, or other runtime execution.

## What

Adds a new CLI target to be used by the `cargo run` command. This CLI target generates an OpenAPI YAML with the following modifications:

1. `/internal` paths are removed. These are not available outside of other in-cluster services.
2. `/namespaces/{namespace}` gets renamed to `/workflows`, as the Tensorlake Ingress design dictates.
3. `/namespaces` path gets removed. These operations (create and list namespaces) are not public.
4. Add security schemes, descriptions, etc.