Check FE health after running each task
If a task failed because FE is unhealthy we learn about this asap
so Server stop schedule tasks on this FE and we reduce overall
latency and rate of failures in the system.

Because this is a hacky implementation we're not checking if the
task failed right now. It's more complecated, will be done later
after refactoring.

Now test_function_process_crash.py from Tensorklake SDK tests
passes even without the 10 sec delay.