Remove test_graph_secrets test
It's the only test that is using function closures which we won't
be able to use once we migrate to ZIP serialized graph code.

Instead of fixing this test I just delete it because it's low
value. It verifies that OSS Executor doesn't implement secrets
and functions that use secretes still can run in OSS. There's
very low probability that this behavoir would ever regress.
That's why removing the test as low value.

Also updated Tensorlake submodule and formatted recently changed code.