Add ExecutorUpdate message to Executor grpc API
This is a mechanism for Executor to report updates to Server so
Server doesn't have to diff these updates out of the existing
ExecutorState. This also allows to not grow ExecutorState for things
that Executor needs to report successfully only once.

Currently ExecutorUpdate includes:

* Completed tasks
* FE startup logs

Also pass real allocation ID to executor in desired state instead of allocation key.
This is to assist with end to end tracing and consistent log tags.