Use DataPayload for InvocationError message/payload
This makes it much more flexible wrt the size of the InvocationError
message and lets SDK own what can be stored there.
Users can now add any size of string message to InvocationError
and in the future we can easily allow DataModels there and etc
if needed.

This also removes the need to store the InvocationError message
in Server state store. All we need is just put DataPayload there.