Track internal state in Function Executor controller
Previously FE Status proto was used for tracking of internal
controller state but FE status proto is it's not sufficient
because we need to also track Terminating state to dedup
FE termination events. So we're splitting reported proto
Status and internal FE controller state in this change.
Then we use the internal state to dedup events.

Also add a lock to FE destroy code to make sure that when
we shutdown FE we don't immediately exit from fe.destroy()
call if another destroy is in progress because fe.destroy()
is also idempotent. When we shutdown FE we need to make sure
that all FE resources were freed on return from FE shutdown.