Backoff properly if an Executor state report fails
We weren't really backing off if Executor had new state report updates
to publish while we were attempting to report a state and it failed.
This is because self._state_report_scheduled_event in Executor state
reporter is set on each 5 sec interval and when there's an update
to publish.

Now we backoff correcly and don't depend on when
self._state_report_scheduled_event is updated.