Log everything happening during grpc channel creation
We had an outage in Dev where Executor stopped re-trying to establish
grpc channel to Server after some time. This is possibly due to a grpc
library call hanging. Adding detailed logs to figure out if this is
the case or not and also to assist future debugging.

Also catch BaseException instead of Exception. Our retry logic could fail
if an unexpected BaseException was raised from grpc lib.

Also bump Indexify and dependencies' versions.