Move static state change building.
Several requests that emit state changes generate them based on information available in the request. Those state changes don't need to be created inside the state store.

This change extracts the step of building state changes for those operations directly in the request, or before the state store is called for writing the request.