Move static state change building. (#1716)
* Move static state change building.

Several requests that emit state changes generate them based on information available in the request. Those state changes don't need to be created inside the state store.

This change extracts the step of building state changes for those operations directly in the request, or before the state store is called for writing the request.

* Update state changes building for UpsertExecutor.

Extract the changes precondition to a different function since they don't need to be part of the write transaction.
Check the changes precondition and create the changes before sending the request to the state store.

* Ensure that the executor state is updated when the hash changed.

* Abstract the creation of UpsertExecutorRequest.

This ensures that state changes are generated properly by not allowing the struct to be initialized directly.