Database abstractions.
This change introduces the beginning of abtracting the current RocksDB logic behind an interface. This will allow us at some point to break the hard dependency of the state with RocksDB, and move the state to an extrernal store.

This change also introduces a new operation in the RocksDB driver to perform CompareAndSwap operations. This is an example of how to implement new operations on top of the driver abstraction.

The state store does not currently use most of this abstraction at the moment. This is my way to prevent any regression issue to land this change. The only change in the current code is the way the RockDB connection is initialized, which should still be compatible with the code in production.