Test the granularity of the epoch data we store.
This change introduces a time mock library that allow us to control time.

With that, we can verify that the times that we store when we create data models have the exact level or granularity that we want.

As an example, if you change the granularity of `TaskBuilder::default_creation_time_ns` from nanoseconds to milliseconds, a test will fail. Before, no test could fail because it was hard for us to know which exact time was generated by `get_epoch_time_in_ns`.