Simplify passing epoch times around functions. (#1740)
* Simplify passing epoch times around functions.

* Revert created_at accuracy.

* Remove EpochTime.

The struct complicates things without providing enough value, and it's messing with time accuracy.

* Run timed based tests in isolation.

This should prevent the tests from failing when the executor is busy.

* Test the granularity of the epoch data we store.

This change introduces a time mock library that allow us to control time.

With that, we can verify that the times that we store when we create data models have the exact level or granularity that we want.

As an example, if you change the granularity of `TaskBuilder::default_creation_time_ns` from nanoseconds to milliseconds, a test will fail. Before, no test could fail because it was hard for us to know which exact time was generated by `get_epoch_time_in_ns`.