Replace the direct usage of RocksDB (#1744)
* Replace the direct usage of TransactionDB.

Instead of passing TransactionDB objects, we're passing RocksDBDriver. This doesn't change much what we have, but it allow us to start putting more logic behing the driver trait.

Repace all usage of `cf_db` with `column_family` to be more explicit of what it is, and also remove the call from within the enum implementation. This makes the code more consistent across the board.

Stop calling `column_family` from within iterators. That call uses a RwLock to access column families. If the lock is held for writing, all read operation wait for it. Instead of calling it for every iteration within loops, we can call it before looping, reducing the lock contention in specific areas.

* Abstract RocksDB transactions away.

With this change, all the specific RocksDB logic lives behind abstractions.

There are still some assumptions in those abstractions, but we can easily modify them as we bring other drivers in.

* Move public functions to traits.

Import the required traits when necessary.