Transaction improvements (#1846)
* Remove transaction lifetimes.

Make the transaction lifetime static. The database should
always outlive transactions, so we can assume this lifetime.
If the database is dropped prematurely, the transaction lifetime
will cause a crash.

* Make RocksDB transaction Send.

Create a Transaction trait that is Send and contains
the functions that we expect a transaction to implement.

This should allow us keep the RocksDB driver even if we have
a new async/await implementation because the TransactionDB for
RocksDB become Send.