Cloud events server (#1865)
* Add cloud events exporter.

The server can be configured to export and read cloud events from
an external source.

The exporter is currently used to send an event when a request is completed.
The exporter runs a background task in charge of sending the payloads to the
external service without blocking the scheduler.

* Remove reader endpoint.

Indexify Server won't read from that endpoint.

* Cleanup invoke routes.

Remove duplicated code to make the stream endpoint easier to understand.

* Propagate all request state change updates.

Send all updates to the cloud events endpoint. That way clients
that connect to that endpoint will receive all the progress updates
coming from the executors and the server.

* Rename CloudEventsExporter to be more precise about what it does.

* Serialize RequestOutcome as an enum.

Downstream services like Inkwell server expect this to follow our conventions.

* Rename RequestOutcome variants to follow the same conventions we have with FunctionRunOutcome.

The enum keys are lowercase. The values for the failure reason are snake_case.

* Add some extra tracing around sending cloud events.

* Process events when one of the two events receivers are waiting.

* Fix event type value.

* Revert version field name change back.

* Graceful shutdown to ensure we export all progress updates.

* Invert select biased.

Ensure that we consume everything in the channel be fore terminating.

* Update Tensorlake package and submodule.