Executor side function call size validation and state report fragmentation
* Set to 4 GB max grpc report state message size so we don't
  get hard failures when sending messages.
* Validate max function call and allocation result sizes.
* Limit size of executor state report messages sent to server
  by fragmenting the messages if they are too big.
  Fragmentation is done by spreading allocation results that
  include large tail function calls into multiple state report
  messages (rpcs).

This ensures that we're sending at least one allocation result
per state report so the system progresses but without sending a
too large state report in one go to not overload Server.

Limit for one state report message size sent to server is 10 MB.

This validation and fragmentation is only implemented for service
stability. FE and local runner side validations are required for
better UX. But it'll take extra time to implement these because
local runner works in a very different way than FE. This is why
this is not implemented yet.