Handle exceptions and send errors back to FE in all possible case
If we don't do this then we fail the alloc on Executor side
and terminate FE instead of giving FE a chance to handle the
error and if it can't do anything about it just fail the alloc
cleanely so the FE is reused for new allocs.

Also remove request metrics logging as we know do this on FE side
as Cloud Events.