Use Idempotency-Key instead of X-Request-Id
This header is a well known HTTP standard. We should follow them
as much as we can because developers might be familiar with it.

As an example, Stripe has been using this standard for years,
so everyone that has ever worked with Stripe is probably familiar
with what an Idempotency-Key is.