Don't use FE's function_call_id as a key in Executor data structures
FE's function_call_id is not going to be unique soon while we're
implementing durability. So Executor needs to stop assuming that
FE provided function_call_id is unique per request.

The change is forward and backward compatible with FE.
Backward compatibility with FE is tested in this PR because
tensorlake SDK version is not updated.
Tensorlake SDK will be updated in a separate PR.